Internal

Vlasiator.MeshInfo — Type

Mesh size information.

Vlasiator.MetaData — Type

Meta data declaration.

Vlasiator.VarInfo — Type

Variable metadata from the vlsv footer, including the unit of the variable as a regular string, the unit of the variable as a LaTeX-formatted string, the description of the variable as a LaTeX-formatted string, and the conversion factor to SI units as a string.

Vlasiator.getSliceCellID — Method

getSliceCellID(meta, slicelocation, maxreflevel; xmin=-Inf, xmax=Inf,

ymin=-Inf, ymax=Inf, zmin=-Inf, zmax=Inf)

Find the cell ids which are needed to plot a 2d cut through of a 3d mesh.

Vlasiator.get_amr_level — Method

get_amr_level(meta::MetaData, cellid)

Return the AMR level of a given cell id.

Vlasiator.get_cell_coordinates — Method

get_cell_coordinates(meta::MetaData, cellid)

Return a given cellid's coordinates.

Vlasiator.get_cell_in_line — Method

get_cell_in_line(meta::MetaData, point1, point2)

Returns cell IDs, distances and coordinates for every cell in a line between two given points. May be improved later with preallocation!

Vlasiator.get_cellid — Method

get_cell_id(meta::MetaData, location)

Return cell ID containing the given spatial location.

Vlasiator.get_max_amr_level — Method

get_max_amr_level(meta::MetaData)

Find the highest refinement level.

Vlasiator.has_parameter — Method

has_parameter(meta::MetaData, param)

Check if vlsv file contains a parameter.

Vlasiator.has_variable — Method

Check if vlsv file contain a variable.

Vlasiator.plot_colormap3dslice — Method

plot_colormap3dslice(meta::MetaData, var (...))

Plot pseudocolor var on a 2D slice of 3D vlsv data.

plot_colormap3dslice(meta, var)

plot_colormap3dslice(meta, var, op="z", origin=1.0, normal="x")

plot_colormap3dslice(data, func, islinear=false)

Vlasiator.plot_pcolormesh — Method

plot_pcolormesh(meta::MetaData, var; op="mag", axisunit="Re", islinear=false)

Plot a 2D pseudocolor var from vlsv.

plot_pcolormesh(meta, var)

plot_pcolormesh(meta, var, axisunit="SI")

plot_pcolormesh(data, func, islinear=false)

Vlasiator.read_mesh — Method

Return mesh related variable.

Vlasiator.read_meta — Method

read_meta(filename; verbose=false)

Return a struct of MetaData from vlsv file.

Vlasiator.read_parameter — Method

read_parameter(meta::MetaData, param)

Return the parameter value from vlsv file.

Vlasiator.read_prep — Method

Return size and type information for the object.

Vlasiator.read_variable — Method

read_variable(meta::MetaData, var)

Return variable value from the vlsv file.

Vlasiator.read_variable_info — Method

read_variable_info(meta::MetaData, var)

Return a struct of VarInfo.

Vlasiator.read_variable_select — Function

read_variable_select(meta::MetaData, var, idIn=UInt[])

Read a variable in a collection of cells.

Vlasiator.read_vector — Method

Return vector data from vlsv file.

Vlasiator.read_xml_footer — Method

Return the xml footer of vlsv.

Vlasiator.refine_data — Method

refine_data(meta, idlist, data, maxreflevel, normal)

Generate scalar data on the finest refinement level.

Vlasiator.set_args — Method

Set plot-related arguments.

Vlasiator.set_plot — Method

Draw customized plot.