Internal

Public APIs

Reader

Vlasiator.MetaData — Type

Meta data declaration.

Vlasiator.VarInfo — Type

Variable metadata from the vlsv footer, including the unit of the variable as a regular string, the unit of the variable as a LaTeX-formatted string, the description of the variable as a LaTeX-formatted string, and the conversion factor to SI units as a string.

Vlasiator.compare — Function

compare(filename1, filename2, tol=1e-4) -> Bool

Check if two VLSV files are approximately identical.

Vlasiator.getamr — Method

getamr(meta, cellid) -> Int

Return the AMR level of a given cell ID.

Vlasiator.getcell — Method

getcell(meta, location) -> Int

Return cell ID containing the given spatial location.

Vlasiator.getcellcoordinates — Method

getcellcoordinates(meta, cellid) -> Vector{Float}

Return a given cell's coordinates.

Vlasiator.getcellinline — Method

getcellinline(meta, point1, point2)

Returns cell IDs, distances and coordinates for every cell in a line between two given points. May be improved later with preallocation!

Vlasiator.getmaxamr — Method

getmaxamr(meta) -> Int

Find the highest refinement level of a given vlsv file.

Vlasiator.getnearestcellwithvdf — Method

Find the nearest spatial cell with f saved of a given cell id.

Vlasiator.getslicecell — Method

getslicecell(meta, slicelocation, maxreflevel; xmin=-Inf, xmax=Inf,
   ymin=-Inf, ymax=Inf, zmin=-Inf, zmax=Inf)

Find the cell ids idlist which are needed to plot a 2d cut through of a 3d mesh, in a direction with non infinity range at slicelocation, and the indexlist, which is a mapping from original order to the cut plane and can be used to select data onto the plane.

Vlasiator.getvcellcoordinates — Method

getvcellcoordinates(meta, vcellids, pop="proton")

Return velocity cells' coordinates of population pop and id vcellids.

Vlasiator.hasname — Method

Check if the XMLElement elem contains a tag with name.

Vlasiator.hasparameter — Method

hasparameter(meta, param) -> Bool

Check if vlsv file contains a parameter.

Vlasiator.hasvariable — Method

Check if the VLSV file contains a variable.

Vlasiator.readmeta — Method

readmeta(filename; verbose=false) -> MetaData

Return MetaData from a vlsv file.

Vlasiator.readparameter — Method

readparameter(meta, param)

Return the parameter value from vlsv file.

Vlasiator.readvariable — Function

readvariable(meta, var, sorted=true) -> Array

Return variable value from the vlsv file. For DCCRG grid, the variables are sorted by cell ID by default.

Vlasiator.readvariable — Method

readvariable(meta, var, idIn) -> Array

Read a variable in a collection of cells.

Vlasiator.readvariableinfo — Method

readvariableinfo(meta, var) -> VarInfo

Return VarInfo about var in the vlsv file linked to meta.

Vlasiator.readvcells — Method

readvcells(meta, cellid; pop="proton")

Read velocity cells from a spatial cell of ID cellid, and return a map of velocity cell ids and corresponding value.

Vlasiator.refinedata — Method

refinedata(meta, idlist, data, maxreflevel, normal) -> Array

Generate scalar data on the finest refinement level.

Vlasiator.showvariables — Method

Display all variables in the VLSV file.

Plotter

Private types and methods

Vlasiator.MeshInfo — Type

Mesh size information.

Vlasiator.getObjInfo — Method

Return size and type information for the object.

Vlasiator.getRotationB — Method

getRotationB(B) -> SMatrix

Obtain a rotation matrix with each column being a unit vector which is parallel (v3) and perpendicular (v1,v2) to the magnetic field B. The two perpendicular directions are chosen based on the reference vector of z-axis in the Cartesian coordinates.

Vlasiator.get_rotation_matrix — Method

get_rotation_matrix(vector, angle)

Creates a rotation matrix that rotates around a unit vector by an angle in radians. References: https://en.wikipedia.org/wiki/Rodrigues'rotationformula https://en.wikipedia.org/wiki/Rotationmatrix#Rotationmatrixfromaxisandangle

Vlasiator.getfooter — Method

Return the xml footer of vlsv.

Vlasiator.readmesh — Method

Return mesh related variable.

Vlasiator.readvector — Method

Return vector data from vlsv file.

Vlasiator.rotateTensorToVectorZ! — Method

rotateTensorToVector(tensor, vector)

Rotates tensor with a rotation matrix that aligns z-axis with vector.

Vlasiator.rotateWithB! — Method

rotateWithB!(T, B)

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB

Vlasiator.rotateWithB — Method

rotateWithB(T, B) -> Matrix

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB!