Internal

Public APIs

Vlasiator.MetaData — Type

Meta data declaration.

Vlasiator.VarInfo — Type

Variable metadata from the vlsv footer.

PyPlot.pcolormesh — Function

pcolormesh(meta::MetaData, var, ax=nothing;
   op=:mag, axisunit=RE, colorscale=Log, vmin=-Inf, vmax=Inf, addcolorbar=true)

Plot a variable using pseudocolor from 2D VLSV data. If ax is provided, then it will plot on that axes.

Optional arguments

op::Symbol: the component of a vector to plot, chosen from :mag, :x, :y, :z.

axisunit::AxisUnit: the unit of axis ∈ RE, SI.

colorscale::ColorScale: whether to use linear scale for data.

vmin::Float: minimum data range. Set to maximum of data if not specified.

vmax::Float: maximum data range. Set to minimum of data if not specified.

addcolorbar::Bool: whether to add a colorbar to the colormesh.

pcolormesh(meta, var)

pcolormesh(meta, var, axisunit=SI)

pcolormesh(data, func, colorscale=Linear)

PyPlot.plot — Method

plot(meta, var; kwargs)

Plot var from meta of 1D VLSV data.

PyPlot.quiver — Method

quiver(meta, var; comp="xy", axisunit=RE, kwargs...)

Wrapper over Matplotlib's quiver function. The comp option can take a subset of "xyz" in any order. axisunit can be chosen from RE, SI. The keyword arguments can be any valid Matplotlib arguments into quiver.

PyPlot.streamplot — Method

streamplot(meta, var; comp="xy", axisunit=RE, kwargs...)

Wrapper over Matplotlib's streamplot function. The comp option can take a subset of "xyz" in any order. axisunit can be chosen from RE, SI. The keyword arguments can be any valid Matplotlib arguments into streamplot.

Vlasiator.compare — Function

compare(filename1, filename2, tol=1e-4) -> Bool

Check if two VLSV files are approximately identical.

Vlasiator.getamr — Method

getamr(meta, cellid) -> Int

Return the AMR level of a given cell ID.

Vlasiator.getcell — Method

getcell(meta, location) -> Int

Return cell ID containing the given spatial location.

Vlasiator.getcellcoordinates — Method

getcellcoordinates(meta, cellid) -> Vector{Float}

Return a given cell's coordinates.

Vlasiator.getcellinline — Method

getcellinline(meta, point1, point2)

Returns cell IDs, distances and coordinates for every cell in a line between two given points. May be improved later with preallocation!

Vlasiator.getmaxamr — Method

getmaxamr(meta) -> Int

Find the highest refinement level of a given vlsv file.

Vlasiator.getnearestcellwithvdf — Method

Find the nearest spatial cell with f saved of a given cell id.

Vlasiator.getslicecell — Method

getslicecell(meta, slicelocation, maxreflevel;
   xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf, zmin=-Inf, zmax=Inf)

Find the cell ids idlist which are needed to plot a 2d cut through of a 3d mesh, in a direction with non infinity range at slicelocation, and the indexlist, which is a mapping from original order to the cut plane and can be used to select data onto the plane.

Vlasiator.getvcellcoordinates — Method

getvcellcoordinates(meta, vcellids, pop="proton")

Return velocity cells' coordinates of population pop and id vcellids.

Vlasiator.hasname — Method

Check if the XMLElement elem contains a tag with name.

Vlasiator.hasparameter — Method

hasparameter(meta, param) -> Bool

Check if vlsv file contains a parameter.

Vlasiator.hasvariable — Method

Check if the VLSV file contains a variable.

Vlasiator.pcolormeshslice — Function

pcolormeshslice(meta, var, ax=nothing (...))

Plot pseudocolor var on a 2D slice of 3D vlsv data. If ax is provided, then it will plot on that axes.

Optional arguments

op::Symbol: the component of a vector to plot, chosen from :mag, :x, :y, :z.

origin::Float: center of slice plane in the normal direction.

normal::Symbol: the normal direction of cut plane, chosen from :x, :y, :z.

axisunit::AxisUnit: the unit of axis ∈ RE, SI.

colorscale::ColorScale: color scale for data ∈ (Linear, Log)

vmin::Float: minimum data range. Set to maximum of data if not specified.

vmax::Float: maximum data range. Set to minimum of data if not specified.

addcolorbar::Bool: whether to add a colorbar to the colormesh.

pcolormeshslice(meta, var)

pcolormeshslice(meta, var, op=:z, origin=1.0, normal=:x)

pcolormeshslice(data, func, colorscale=Log)

Vlasiator.plot_vdf — Method

plot_vdf(meta, location; kwargs...)

Plot the 2D slice cut of phase space distribution function at location within velocity range limits.

Optional arguments

limits: velocity space range given in [xmin, xmax, ymin, ymax].

slicetype: string for choosing the slice type from "xy", "xz", "yz", "bperp", "bpar",

"bpar1".

center: string for setting the reference frame from "bulk", "peak".

vslicethick: setting the velocity space slice thickness in the normal direction. If set

to 0, the whole distribution along the normal direction is projected onto a plane. Currently this is only meaningful when center is set such that a range near the bulk/peak normal velocity is selected!

weight: symbol for choosing distribution weights from phase space density or particle

flux.

Vlasiator.readmeta — Method

readmeta(filename; verbose=false) -> MetaData

Return MetaData from a vlsv file.

Vlasiator.readparameter — Method

readparameter(meta, param)

Return the parameter value from vlsv file.

Vlasiator.readvariable — Function

readvariable(meta, var, sorted=true) -> Array

Return variable value from the vlsv file. By default sorted=true, which means that for DCCRG grid the variables are sorted by cell ID.

Vlasiator.readvariable — Method

readvariable(meta, var, idIn) -> Array

Read a variable in a collection of cells.

Vlasiator.readvariableinfo — Method

readvariableinfo(meta, var) -> VarInfo

Return VarInfo about var in the vlsv file linked to meta.

Vlasiator.readvcells — Method

readvcells(meta, cellid; pop="proton")

Read velocity cells from a spatial cell of ID cellid, and return a map of velocity cell ids and corresponding value.

Vlasiator.refinedata — Method

refinedata(meta, idlist, data, maxreflevel, normal) -> Array

Generate scalar data on the finest refinement level.

Vlasiator.showvariables — Method

Display all variables in the VLSV file.

PyPlot helpers

Vlasiator.PlotArgs — Type

Plotting arguments.

Vlasiator.plot_prep2d — Method

Generate axis and data for 2D plotting.

Vlasiator.set_args — Method

Set plot-related arguments.

Vlasiator.set_colorbar — Method

Set colorbar norm and ticks.

Vlasiator.set_plot — Method

Configure customized plot.

Private APIs

Vlasiator.AxisUnit — Type

Axis unit type. Currently supported: SI, RE.

Vlasiator.ColorScale — Type

Color scales type for 2D plots. Currently supported: Log, Linear.

Vlasiator.MeshInfo — Type

Mesh size information.

Vlasiator.getObjInfo — Method

Return size and type information for the object.

Vlasiator.getRotationB — Method

getRotationB(B) -> SMatrix

Obtain a rotation matrix with each column being a unit vector which is parallel (v3) and perpendicular (v1,v2) to the magnetic field B. The two perpendicular directions are chosen based on the reference vector of z-axis in the Cartesian coordinates.

Vlasiator.get_rotation_matrix — Method

get_rotation_matrix(vector, angle)

Creates a rotation matrix that rotates around a unit vector by an angle in radians. References: https://en.wikipedia.org/wiki/Rodrigues'rotationformula https://en.wikipedia.org/wiki/Rotationmatrix#Rotationmatrixfromaxisandangle

Vlasiator.getfooter — Method

Return the xml footer of vlsv.

Vlasiator.readmesh — Method

Return mesh related variable.

Vlasiator.readvector — Method

Return vector data from vlsv file.

Vlasiator.rotateTensorToVectorZ! — Method

rotateTensorToVector(tensor, vector)

Rotates tensor with a rotation matrix that aligns z-axis with vector.

Vlasiator.rotateWithB! — Method

rotateWithB!(T, B)

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB

Vlasiator.rotateWithB — Method

rotateWithB(T, B) -> Matrix

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB!