Internal

Public APIs

Vlasiator.MetaData — Type

VLSV meta data.

Vlasiator.VarInfo — Type

Variable metadata from the vlsv footer.

Base.ndims — Method

ndims(meta) -> Int

Return the dimension of VLSV data.

Vlasiator.compare — Function

compare(filename1, filename2, tol=1e-4) -> Bool

Check if two VLSV files are approximately identical.

Vlasiator.getcell — Method

getcell(meta, location) -> Int

Return cell ID containing the given spatial location, excluding domain boundaries.

Vlasiator.getcellcoordinates — Method

getcellcoordinates(meta, cellid) -> Vector{Float}

Return a given cell's coordinates.

Vlasiator.getcellinline — Method

getcellinline(meta, point1, point2) -> cellids, distances, coords

Returns cell IDs, distances and coordinates for every cell in a line between two given points point1 and point2. May be improved later with preallocation!

Vlasiator.getchildren — Method

getchildren(meta, cellid) -> Vector{Int}

Return direct children of cellid.

Vlasiator.getlevel — Method

getlevel(meta, cellid) -> Int

Return the AMR level of a given cell ID. Note that this function does not check if the VLSV file of meta actually contains cellid: it may be shadowed by refined children.

Vlasiator.getnearestcellwithvdf — Method

getnearestcellwithvdf(meta, id) -> Int

Find the nearest spatial cell with VDF saved of a given cell id in the file meta.

Vlasiator.getparent — Method

getparent(meta, cellid) -> Int

Return the parent cell ID of given child cellid.

Vlasiator.getsiblings — Method

getsiblings(meta, cellid) -> Vector{Int}

Return sibling cells of a given cellid, including itself.

Vlasiator.getslicecell — Method

getslicecell(meta, slicelocation;
   xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf, zmin=-Inf, zmax=Inf) -> idlist, indexlist

Find the cell ids idlist which are needed to plot a 2d cut through of a 3d mesh, in a direction with non infinity range at slicelocation, and the indexlist, which is a mapping from original order to the cut plane and can be used to select data onto the plane.

Vlasiator.getvcellcoordinates — Method

getvcellcoordinates(meta, vcellids, pop="proton")

Return velocity cells' coordinates of population pop and id vcellids.

Vlasiator.hasname — Method

Check if the XMLElement elem contains a tag with name.

Vlasiator.hasparameter — Method

hasparameter(meta, param) -> Bool

Check if vlsv file contains a parameter.

Vlasiator.hasvariable — Method

hasvariable(meta, var) -> Bool

Check if the VLSV file contains a variable.

Vlasiator.hasvdf — Method

hasvdf(meta) -> Bool

Check if VLSV file contains VDF.

Vlasiator.isparent — Method

isparent(meta, cellid) -> Bool

Check if cellid is a parent cell.

Vlasiator.readlog — Method

readlog(file)

Read the run log file, check the iteration status and return the timestamps (exluding the last) as well as the model running speed in physical seconds per simulated seconds.

Vlasiator.readmeta — Method

readmeta(filename; verbose=false) -> MetaData

Return MetaData from a vlsv file.

Vlasiator.readparameter — Method

readparameter(meta, param)

Return the parameter value from vlsv file.

Vlasiator.readvariable — Function

readvariable(meta, var, sorted=true) -> Array

Return variable value from the vlsv file. By default sorted=true, which means that for DCCRG grid the variables are sorted by cell ID.

Vlasiator.readvariable — Method

readvariable(meta, var, ids) -> Array

Read a variable var in a collection of cells ids.

Vlasiator.readvariablemeta — Method

readvariablemeta(meta, var) -> VarInfo

Return VarInfo about var in the vlsv file linked to meta.

Vlasiator.readvcells — Method

readvcells(meta, cellid; pop="proton")

Read velocity cells from a spatial cell of ID cellid, and return a map of velocity cell ids and corresponding value.

Vlasiator.refineslice — Method

refineslice(meta, idlist, data, normal) -> Array

Generate scalar data on the finest refinement level given cellids idlist and variable data on the slice perpendicular to normal.

Vlasiator.write_vtk — Method

write_vtk(meta::MetaData; kwargs...)
write_vtk(filename; kwargs...)

Convert VLSV file to VTK format.

Keyword arguments

vars=[""]: select which variables to convert.

ascii=false: output stored in ASCII or compressed binary format.

vti=false: generate image files on the highest refinement level only.

verbose=false: display logs during conversion.

PyPlot helpers

Private APIs

Vlasiator.AxisUnit — Type

Axis unit type. Currently supported: SI, RE.

Vlasiator.ColorScale — Type

Color scales type for 2D plots. Currently supported: Log, Linear.

Vlasiator.VMeshInfo — Type

Velocity mesh information.

Vlasiator.fillmesh — Method

fillmesh(meta::MetaData, vars; verbose=false)

Fill the DCCRG mesh with quantity of vars on all refinement levels.

Return arguments

celldata::Vector{Vector{Array}}: data for each variable on each AMR level.

vtkGhostType::Array{UInt8}: cell status (to be completed!).

Vlasiator.get1stcell — Method

Return the first cellid - 1 on my level.

Vlasiator.getObjInfo — Method

Return size and type information for the object.

Vlasiator.getRotationB — Method

getRotationB(B) -> SMatrix

Obtain a rotation matrix with each column being a unit vector which is parallel (v3) and perpendicular (v1,v2) to the magnetic field B. The two perpendicular directions are chosen based on the reference vector of z-axis in the Cartesian coordinates.

Vlasiator.get_rotation_matrix — Method

get_rotation_matrix(vector, angle)

Creates a rotation matrix that rotates around a unit vector by an angle in radians. References: https://en.wikipedia.org/wiki/Rodrigues'rotationformula https://en.wikipedia.org/wiki/Rotationmatrix#Rotationmatrixfromaxisandangle

Vlasiator.getfooter — Method

Return the xml footer of vlsv.

Vlasiator.getindexes — Method

Compute every cell id's x, y and z indexes on the given refinement level (0-based).

Vlasiator.readmesh — Method

Return mesh related variable.

Vlasiator.readvector — Method

Return vector data from vlsv file.

Vlasiator.rotateTensorToVectorZ! — Method

rotateTensorToVector(tensor, vector)

Rotates tensor with a rotation matrix that aligns z-axis with vector.

Vlasiator.rotateWithB! — Method

rotateWithB!(T, B)

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB

Vlasiator.rotateWithB — Method

rotateWithB(T, B) -> Matrix

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB!

Vlasiator.save_image — Function

save_image(meta::MetaData, file, vars, data, vtkGhostType, level,
   ascii=false, append=true)

Save data of name vars at AMR level into VTK image file of name file.

Arguments

file::String: output file name.

vars::Vector{String}: variable names to be saved.

data::Vector{Vector}: data for all the variables on each refinement level.

vtkGhostType::Array{UInt8}: array for visibility control.

level::Int: refinement level (0-based).

ascii=false: save output in ASCII or binary format.

append=true: determines whether to append data at the end of file or do in-block writing.