Internal

Public APIs

Vlasiator.MetaVLSV — Type

VLSV meta data.

Vlasiator.VarInfo — Type

Variable information from the VLSV footer.

PyPlot.pcolormesh — Function

pcolormesh(meta::MetaVLSV, var::AbstractString, ax=nothing;
   op=:mag, axisunit=RE, colorscale=Linear, vmin=-Inf, vmax=Inf, addcolorbar=true,
   kwargs...)

Plot a variable using pseudocolor from 2D VLSV data. If ax is provided, then it will plot on that axes. If 3D or AMR grid detected, it will pass arguments to pcolormeshslice.

Optional arguments

op::Symbol: the component of a vector, chosen from :mag, :x, :y, :z, :1, :2, :3.

axisunit::AxisUnit: the unit of axis ∈ RE, SI.

colorscale::ColorScale: Linear, Log, or SymLog.

vmin::Float: minimum data range. Set to maximum of data if not specified.

vmax::Float: maximum data range. Set to minimum of data if not specified.

addcolorbar::Bool: whether to add a colorbar to the colormesh.

pcolormesh(meta, var)

pcolormesh(meta, var, axisunit=SI)

pcolormesh(data, func, colorscale=Log)

PyPlot.plot — Function

plot(meta, var, ax=nothing; kwargs)

Plot var from meta of 1D VLSV data. If ax===nothing, plot on the current active axes.

PyPlot.quiver — Function

quiver(meta, var, ax=nothing; comp="xy", axisunit=RE, stride=10, kwargs...)

Wrapper over Matplotlib's quiver function. If ax===nothing, plot on the current active axes. The comp option can take a subset of "xyz" in any order. axisunit can be chosen from RE, SI. The keyword arguments can be any valid Matplotlib arguments into quiver.

Optional arguments

comp: a subset of "xyz" in any order.

axisunit: chosen from RE and SI.

stride::Integer: arrow strides in number of cells.

PyPlot.streamplot — Function

streamplot(meta, var, ax=nothing; comp="xy", axisunit=RE, kwargs...)

Wrapper over Matplotlib's streamplot function. The comp option can take a subset of "xyz" in any order. axisunit can be chosen from RE, SI. The keyword arguments can be any valid Matplotlib arguments into streamplot.

Optional arguments

comp: a subset of "xyz" in any order.

axisunit: chosen from RE and SI.

Vlasiator.getcell — Method

getcell(meta, location) -> UInt

Return cell ID containing the given spatial location, excluding domain boundaries. Only accept 3D location.

Vlasiator.getcellcoordinates — Method

getcellcoordinates(meta, cid)

Return a given cell's coordinates.

Vlasiator.getcellinline — Method

getcellinline(meta, point1, point2) -> cellids, distances, coords

Returns cell IDs, distances and coordinates for every cell in a line between two given points point1 and point2. TODO: preallocation?

Vlasiator.getcellwithvdf — Method

getcellwithvdf(meta) -> cellids

Get all the cell IDs with VDF saved associated with meta.

Vlasiator.getchildren — Method

getchildren(meta, cid) -> Vector{Int}

Return direct children of cid.

Vlasiator.getdensity — Method

getdensity(meta, VDF; species="proton")
getdensity(meta, vcellids, vcellf; species="proton")

Get density from VDF of species associated with meta, n = ∫ f(r,v) dV. Alternatively, one can pass vcellids as local indices of nonzero VDFs and vcellf as their corresponding values.

Vlasiator.getlevel — Method

getlevel(meta, cid) -> Int

Return the AMR level of a given cell ID cid.

Warning

This function does not check if the VLSV file of meta actually contains cid; it may be shadowed by refined children.

Vlasiator.getmaxwellianity — Method

getmaxwellianity(meta, VDF; species="proton")

Obtain the Maxwellian similarity factor -log(1/(2n) * ∫ |f - g| dv), where f is the VDF from Vlasiator and g is the analytical Maxwellian distribution that generates the same density as f. The value ranges from [0, +∞], with 0 meaning not Maxwellian-distributed at all, and +∞ a perfect Maxwellian distribution.

Vlasiator.getnearestcellwithvdf — Method

getnearestcellwithvdf(meta, id) -> UInt

Find the nearest spatial cell with VDF saved of a given cell id associated with meta.

Vlasiator.getparent — Method

getparent(meta, cid) -> Int

Return the parent cell ID of given child cid.

Vlasiator.getpressure — Method

getpressure(meta, VDF; species="proton")

Get pressure tensor (6 components) of species from VDF associated with meta, pᵢⱼ = m/3 * ∫ (v - u)ᵢ(v - u)ⱼ * f(r,v) dV.

Vlasiator.getsiblings — Method

getsiblings(meta, cid) -> Vector{Int}

Return sibling cells of a given cid, including itself.

Vlasiator.getslicecell — Method

getslicecell(meta, sliceoffset, idim, minCoord, maxCoord) -> idlist, indexlist

Find the cell IDs idlist which are needed to plot a 2d cut through of a 3d mesh, in a direction idim at sliceoffset, and the indexlist, which is a mapping from original order to the cut plane and can be used to select data onto the plane.

Vlasiator.getvcellcoordinates — Method

getvcellcoordinates(meta, vcellids, species="proton")

Return velocity cells' coordinates of species and vcellids.

Vlasiator.getvelocity — Method

getvelocity(meta, VDF; species="proton")
getvelocity(meta, vcellids, vcellf; species="proton")

Get bulk velocity from VDF of species, u = ∫ v * f(r,v) dV / n. Alternatively, one can pass vcellids, vcellf, as in getdensity.

Vlasiator.hasname — Method

Check if the XML element contains a tag with name.

Vlasiator.hasparameter — Method

hasparameter(meta, param) -> Bool

Check if the VLSV file contains a certain parameter param.

Vlasiator.hasvariable — Method

hasvariable(meta, var) -> Bool

Check if the VLSV file associated with meta contains a variable var.

Vlasiator.isparent — Method

isparent(meta, cid) -> Bool

Check if cid is a parent cell.

Vlasiator.issame — Function

issame(file1, file2, tol=1e-4; verbose=false) -> Bool

Check if two VLSV files file1 and file2 are approximately identical, under relative tolerance tol.

Vlasiator.load — Method

load(file) -> MetaVLSV

Generate a MetaVLSV object from file of VLSV format.

Vlasiator.pcolormeshslice — Function

pcolormeshslice(meta, var, ax=nothing; kwargs...)

Plot pseudocolor var on a 2D slice of 3D vlsv data. If ax is provided, then it will plot on that axes. It would be easier to call pcolormesh, since it auto-detects dimension.

Optional arguments

op::Symbol: the component of a vector, chosen from :mag, :x, :y, :z, :1, :2, :3.

origin::Float: center of slice plane in the normal direction.

normal::Symbol: the normal direction of cut plane, chosen from :x, :y, :z.

axisunit::AxisUnit: the unit of axis ∈ RE, SI.

colorscale::ColorScale: color scale for data ∈ (Linear, Log, SymLog)

vmin::Real: minimum data range. Set to maximum of data if not specified.

vmax::Real: maximum data range. Set to minimum of data if not specified.

addcolorbar::Bool: whether to add a colorbar to the colormesh.

pcolormeshslice(meta, var)

pcolormeshslice(meta, var, op=:z, origin=1.0, normal=:x)

pcolormeshslice(data, func, colorscale=Log)

Vlasiator.plotmesh — Function

plotmesh(meta; projection="3d", origin=0.0, marker="+", kwargs...)

Plot mesh cell centers from axis view projection. projection should be either "3d", "x", "y" or "z". origin is center of projection plane in the normal direction.

Vlasiator.readlog — Method

readlog(file)

Read the run log file, check the iteration status and return the timestamps (exluding the last) as well as the model running speed in physical seconds per simulated seconds.

Vlasiator.readparameter — Method

readparameter(meta, param)

Return the parameter value from the VLSV file associated with meta.

Vlasiator.readvariable — Function

readvariable(meta::MetaVLSV, var, sorted::Bool=true) -> Array

Return variable value of var from the VLSV file associated with meta. By default sorted=true, which means that for DCCRG grid the variables are sorted by cell ID.

Vlasiator.readvariable — Method

readvariable(meta::MetaVLSV, var, ids) -> Array

Read variable var in a collection of cells ids associated with meta.

Vlasiator.readvariablemeta — Method

readvariablemeta(meta, var) -> VarInfo

Return VarInfo about var in the VLSV file associated with meta.

Vlasiator.readvcells — Method

readvcells(meta, cid; species="proton") -> vcellids, vcellf

Read velocity cells of species from a spatial cell of ID cid associated with meta, and return a map of velocity cell ids vcellids and corresponding value vcellf.

Vlasiator.refineslice — Method

refineslice(meta, idlist, data, normal) -> Array

Generate scalar data on the finest refinement level given cellids idlist and variable data on the slice perpendicular to normal.

Vlasiator.vdfslice — Function

vdfslice(meta, location, ax=nothing; kwargs...)

Plot the 2D slice cut of phase space distribution function at location within velocity range limits. If ax===nothing, plot on the current active axes.

Optional arguments

unit::AxisUnit: location unit in SI, RE.

unitv::String: velocity unit in ("km/s", "m/s").

limits::Vector{Real}: velocity space range given in [xmin, xmax, ymin, ymax].

slicetype: symbol for choosing the slice type from :xy, :xz, :yz, :bperp, :bpar, :bpar1.

center: symbol for setting the reference frame from :bulk, :peak.

vslicethick: setting the velocity space slice thickness in the normal direction. If set

to 0, the whole distribution along the normal direction is projected onto a plane. Currently this is only meaningful when center is set such that a range near the bulk/peak normal velocity is selected!

fmin, fmax: minimum and maximum VDF values for plotting.

weight::Symbol: choosing distribution weights from phase space density or particle flux

between :particle and :flux.

flimit: minimum VDF threshold for plotting.

kwargs...: any valid keyword argument for hist2d.

Vlasiator.write_vlsv — Method

write_vlsv(filein, fileout, newvars::Vector{Tuple{Vector, String, VarInfo}};
   force=false)

Generate a new VLSV fileout based on filein, with newvars added. force=true overwrites the existing fileout.

Vlasiator.write_vtk — Method

write_vtk(meta::MetaVLSV; kwargs...)
write_vtk(file; kwargs...)

Convert VLSV file to VTK format.

Keyword arguments

vars=[""]: select which variables to convert.

ascii=false: output stored in ASCII or compressed binary format.

maxamronly=false: generate image files on the highest refinement level only.

verbose=false: display logs during conversion.

PyPlot helpers

Vlasiator.set_colorbar — Function

set_colorbar(colorscale::ColorScale, v1, v2, data=[1.0]) -> cnorm, cticks

Set colorbar norm and ticks in a given range v1 to v2 for data in colorscale. Matplotlib's Colormap Normalization Section presents various kinds of normlizations beyond linear, logarithmic and symmetric logarithmic, like centered, discrete, and two slope norm. For fine-grain control, it is suggested to use the norm methods from matplotlib.colors. For instance,

julia> cnorm = matplotlib.colors.CenteredNorm(); # after matplotlib v3.4
julia> cnorm = matplotlib.colors.BoundaryNorm(boundaries=[0, 1], ncolors=2);
julia> cnorm = matplotlib.colors.TwoSlopeNorm(vmin=-500., vcenter=0, vmax=4000);

There are also various options for the ticks available in matplotlib.ticker.

Vlasiator.set_plot — Method

Configure customized plot.

Private APIs

Vlasiator.AxisUnit — Type

Axis unit type. Currently supported: SI, RE.

Vlasiator.ColorScale — Type

Color scales type for 2D plots. Currently supported: Log, Linear, SymLog.

Vlasiator.PlotArgs — Type

Plotting arguments.

Vlasiator.VMeshInfo — Type

Velocity mesh information.

Base.ndims — Method

ndims(meta) -> Int

Return the simulation spatial dimension of VLSV data.

Base.size — Method

File size in bytes.

Vlasiator.curl — Method

curl(dx, A::AbstractArray)

Calculate 2nd order cell-centered ∇×A where A is a 4D array of size (3, nx, ny, nz) and dx is a vector of grid intervals in each dimension.

Vlasiator.divergence — Method

divergence(dx, A::AbstractArray)

Calculate 2nd order cell-centered ∇⋅A where A is a 4D array of size (3, nx, ny, nz) and dx is a vector of grid intervals in each dimension.

Vlasiator.fillmesh — Method

fillmesh(meta::MetaVLSV, vars; verbose=false) -> celldata, vtkGhostType

Fill the DCCRG mesh with quantity of vars on all refinement levels.

Return arguments

celldata::Vector{Vector{Array}}: data for each variable on each AMR level.

vtkGhostType::Array{UInt8}: cell status (to be completed!).

Vlasiator.flatten — Method

flatten(vmesh::VMeshInfo, vcellids, vcellf)

Flatten vblock-organized VDFs into x–>y–>z ordered 3D VDFs. vcellids are local indices of nonzero VDFs and vcellf are their corresponding values.

Vlasiator.get1stcell — Method

Return the first cellid - 1 on mylevel given ncells on this level.

Vlasiator.getObjInfo — Method

Return size and type information for the inquired parameter/data with name.

Vlasiator.getRotationMatrix — Method

getRotationMatrix(axis, angle) --> SMatrix{3,3}

Create a rotation matrix for rotating a 3D vector around a unit axis by an angle in radians. Reference: https://en.wikipedia.org/wiki/Rotationmatrix#Rotationmatrixfromaxisandangle

Vlasiator.get_axis — Method

Return x and y ranges for 2D.

Vlasiator.getdata2d — Method

Return 2d scalar/vector data. Nonpublic since it won't work with DCCRG AMR.

Vlasiator.getfooter — Method

Return the XML footer of opened VLSV file.

Vlasiator.getindexes — Method

Compute every cell id's x, y and z indexes on the given refinement level (0-based).

Vlasiator.gradient — Method

gradient(dx, A::AbstractArray)

Calculate 2nd order cell-centered ∇A where A is a 3D scalar array of size (nx, ny, nz) and dx is a vector of grid intervals in each dimension.

Vlasiator.prep2d — Function

prep2d(meta, var, op=:none) -> Array

Obtain data from meta of var for 2D plotting. Use op to select vector components.

Vlasiator.prep2dslice — Method

prep2dslice(meta::MetaVLSV, var, normal, op, pArgs::PlotArgs)

Return data of var on a uniform 2D mesh on the finest AMR level. Use normal to select the plane orientation, and op to select the component of a vector, same as in pcolormeshslice.

Vlasiator.prep_vdf — Method

prep_vdf(meta::MetaVLSV, location; kwargs...)

Return the cell velocities v1, v2, bin ranges r1, r2, cell VDF values fweight, and strings of labels and titles for VDF plots.

Optional arguments

unit::AxisUnit: location unit in SI, RE.

unitv::String: velocity unit in ("km/s", "m/s").

limits::Vector{Real}: velocity space range given in [xmin, xmax, ymin, ymax].

slicetype: symbol for choosing the slice type from :xy, :xz, :yz, :bperp, :bpar, :bpar1.

center: symbol for setting the reference frame from :bulk, :peak.

vslicethick: setting the velocity space slice thickness in the normal direction. If set

to 0, the whole distribution along the normal direction is projected onto a plane. Currently this is only meaningful when center is set such that a range near the bulk/peak normal velocity is selected!

weight::Symbol: choosing distribution weights from phase space density or particle flux

between :particle and :flux.

flimit: minimum VDF threshold for plotting.

verbose: display the selection process.

Vlasiator.readmesh — Method

Return mesh related variable.

Vlasiator.readvector — Method

Return vector of name from the VLSV file with footer associated with stream fid.

Vlasiator.rotateTensorToVectorZ — Method

rotateTensorToVectorZ(tensor, vector)

Rotate tensor with a rotation matrix that aligns the 3rd direction with vector, which is equivalent to change the basis from (i,j,k) to (i′,j′,k′) where k′ ∥ vector. Reference: https://math.stackexchange.com/questions/2303869/tensor-rotation

Vlasiator.save_image — Function

save_image(meta::MetaVLSV, file, vars, data, vtkGhostType, level,
   ascii=false, append=true)

Save data of name vars at AMR level into VTK image file of name file.

Arguments

file::String: output file name.

vars::Vector{String}: variable names to be saved.

data::Vector{Vector}: data for all the variables on each refinement level.

vtkGhostType::Array{UInt8}: array for visibility control.

level::Int: refinement level (0-based).

ascii=false: save output in ASCII or binary format.

append=true: determines whether to append data at the end of file or do in-block writing.

Vlasiator.set_args — Method

set_args(meta::MetaVLSV, var, axisunit::AxisUnit; normal::Symbol=:none, origin=0.0)

Set plot-related arguments of var in axisunit. normal and origin are used for 2D slices of 3D data, as specified in pcolormeshslice.

Vlasiator.set_lim — Function

set_lim(vmin, vmax, data, colorscale=Linear)

Set colormap limits vmin, vmax for data under scale colorscale.