Internal

Public APIs

Vlasiator.MetaVLSV — Type

VLSV meta data.

Vlasiator.VarInfo — Type

Variable MetaVLSV from the vlsv footer.

PyPlot.pcolormesh — Function

pcolormesh(meta::MetaVLSV, var::AbstractString, ax=nothing;
   op=:mag, axisunit=RE, colorscale=Linear, vmin=-Inf, vmax=Inf, addcolorbar=true,
   kwargs...)

Plot a variable using pseudocolor from 2D VLSV data. If ax is provided, then it will plot on that axes. If 3D or AMR grid detected, it will pass arguments to pcolormeshslice.

Optional arguments

op::Symbol: the component of a vector, chosen from :mag, :x, :y, :z, :1, :2, :3.

axisunit::AxisUnit: the unit of axis ∈ RE, SI.

colorscale::ColorScale: Linear, Log, or SymLog.

vmin::Float: minimum data range. Set to maximum of data if not specified.

vmax::Float: maximum data range. Set to minimum of data if not specified.

addcolorbar::Bool: whether to add a colorbar to the colormesh.

pcolormesh(meta, var)

pcolormesh(meta, var, axisunit=SI)

pcolormesh(data, func, colorscale=Log)

PyPlot.plot — Function

plot(meta, var, ax=nothing; kwargs)

Plot var from meta of 1D VLSV data. If ax===nothing, plot on the current active axes.

PyPlot.quiver — Function

quiver(meta, var, ax=nothing; comp="xy", axisunit=RE, stride=10, kwargs...)

Wrapper over Matplotlib's quiver function. If ax===nothing, plot on the current active axes. The comp option can take a subset of "xyz" in any order. axisunit can be chosen from RE, SI. The keyword arguments can be any valid Matplotlib arguments into quiver.

Optional arguments

comp: a subset of "xyz" in any order.

axisunit: chosen from RE and SI.

stride::Integer: arrow strides in number of cells.

PyPlot.streamplot — Function

streamplot(meta, var, ax=nothing; comp="xy", axisunit=RE, kwargs...)

Wrapper over Matplotlib's streamplot function. The comp option can take a subset of "xyz" in any order. axisunit can be chosen from RE, SI. The keyword arguments can be any valid Matplotlib arguments into streamplot.

Optional arguments

comp: a subset of "xyz" in any order.

axisunit: chosen from RE and SI.

Vlasiator.getcell — Method

getcell(meta, location) -> UInt

Return cell ID containing the given spatial location, excluding domain boundaries. Only accept 3D location.

Vlasiator.getcellcoordinates — Method

getcellcoordinates(meta, cid) -> Vector{Float}

Return a given cell's coordinates.

Vlasiator.getcellinline — Method

getcellinline(meta, point1, point2) -> cellids, distances, coords

Returns cell IDs, distances and coordinates for every cell in a line between two given points point1 and point2. May be improved later with preallocation!

Vlasiator.getcellwithvdf — Method

getcellwithvdf(meta) -> cellids

Get all the cell IDs with VDF saved.

Vlasiator.getchildren — Method

getchildren(meta, cid) -> Vector{Int}

Return direct children of cid.

Vlasiator.getlevel — Method

getlevel(meta, cid) -> Int

Return the AMR level of a given cell ID. Note that this function does not check if the VLSV file of meta actually contains cid: it may be shadowed by refined children.

Vlasiator.getnearestcellwithvdf — Method

getnearestcellwithvdf(meta, id) -> Int

Find the nearest spatial cell with VDF saved of a given cell id in the file meta.

Vlasiator.getparent — Method

getparent(meta, cid) -> Int

Return the parent cell ID of given child cid.

Vlasiator.getsiblings — Method

getsiblings(meta, cid) -> Vector{Int}

Return sibling cells of a given cid, including itself.

Vlasiator.getslicecell — Method

getslicecell(meta, sliceoffset, idim, minCoord, maxCoord) -> idlist, indexlist

Find the cell ids idlist which are needed to plot a 2d cut through of a 3d mesh, in a direction idim at sliceoffset, and the indexlist, which is a mapping from original order to the cut plane and can be used to select data onto the plane.

Vlasiator.getvcellcoordinates — Method

getvcellcoordinates(meta, vcellids, pop="proton")

Return velocity cells' coordinates of population pop and id vcellids.

Vlasiator.hasname — Method

Check if the XML element contains a tag with name.

Vlasiator.hasparameter — Method

hasparameter(meta, param) -> Bool

Check if the vlsv file contains a certain parameter.

Vlasiator.hasvariable — Method

hasvariable(meta, var) -> Bool

Check if the VLSV file contains a variable.

Vlasiator.isparent — Method

isparent(meta, cid) -> Bool

Check if cid is a parent cell.

Vlasiator.issame — Function

issame(file1, file2, tol=1e-4; verbose=false) -> Bool

Check if two VLSV files are approximately identical.

Vlasiator.load — Method

load(file) -> MetaVLSV

Return MetaVLSV from a vlsv file.

Vlasiator.pcolormeshslice — Function

pcolormeshslice(meta, var, ax=nothing; kwargs...)

Plot pseudocolor var on a 2D slice of 3D vlsv data. If ax is provided, then it will plot on that axes. It would be easier to call pcolormesh, since it auto-detects dimension.

Optional arguments

op::Symbol: the component of a vector, chosen from :mag, :x, :y, :z, :1, :2, :3.

origin::Float: center of slice plane in the normal direction.

normal::Symbol: the normal direction of cut plane, chosen from :x, :y, :z.

axisunit::AxisUnit: the unit of axis ∈ RE, SI.

colorscale::ColorScale: color scale for data ∈ (Linear, Log, SymLog)

vmin::Real: minimum data range. Set to maximum of data if not specified.

vmax::Real: maximum data range. Set to minimum of data if not specified.

addcolorbar::Bool: whether to add a colorbar to the colormesh.

pcolormeshslice(meta, var)

pcolormeshslice(meta, var, op=:z, origin=1.0, normal=:x)

pcolormeshslice(data, func, colorscale=Log)

Vlasiator.plot_vdf — Function

plot_vdf(meta, location, ax=nothing; kwargs...)

Plot the 2D slice cut of phase space distribution function at location within velocity range limits. If ax===nothing, plot on the current active axes.

Optional arguments

unit::AxisUnit: axis unit in SI, RE.

limits::Vector{Real}: velocity space range given in [xmin, xmax, ymin, ymax].

slicetype: symbol for choosing the slice type from :xy, :xz, :yz, :bperp, :bpar, :bpar1.

center: symbol for setting the reference frame from :bulk, :peak.

vslicethick: setting the velocity space slice thickness in the normal direction. If set

to 0, the whole distribution along the normal direction is projected onto a plane. Currently this is only meaningful when center is set such that a range near the bulk/peak normal velocity is selected!

weight::Symbol: choosing distribution weights from phase space density or particle flux

between :particle and :flux.

kwargs...: any valid keyword argument for hist2d.

Vlasiator.plotmesh — Function

plotmesh(meta; projection="3d", origin=0.0, marker="+", kwargs...)

Plot mesh cell centers from axis view projection. projection should be either "3d", "x", "y" or "z". origin is center of projection plane in the normal direction.

Vlasiator.readlog — Method

readlog(file)

Read the run log file, check the iteration status and return the timestamps (exluding the last) as well as the model running speed in physical seconds per simulated seconds.

Vlasiator.readparameter — Method

readparameter(meta, param)

Return the parameter value from vlsv file.

Vlasiator.readvariable — Function

readvariable(meta::MetaVLSV, var, sorted::Bool=true) -> Array

Return variable value of var from the vlsv file. By default sorted=true, which means that for DCCRG grid the variables are sorted by cell ID.

Vlasiator.readvariable — Method

readvariable(meta::MetaVLSV, var, ids) -> Array

Read a variable var in a collection of cells ids.

Vlasiator.readvariablemeta — Method

readvariablemeta(meta, var) -> VarInfo

Return VarInfo about var in the vlsv file linked to meta.

Vlasiator.readvcells — Method

readvcells(meta, cid; pop="proton")

Read velocity cells from a spatial cell of ID cid, and return a map of velocity cell ids and corresponding value.

Vlasiator.refineslice — Method

refineslice(meta, idlist, data, normal) -> Array

Generate scalar data on the finest refinement level given cellids idlist and variable data on the slice perpendicular to normal.

Vlasiator.write_vtk — Method

write_vtk(meta::MetaVLSV; kwargs...)
write_vtk(file; kwargs...)

Convert VLSV file to VTK format.

Keyword arguments

vars=[""]: select which variables to convert.

ascii=false: output stored in ASCII or compressed binary format.

maxamronly=false: generate image files on the highest refinement level only.

verbose=false: display logs during conversion.

PyPlot helpers

Vlasiator.set_colorbar — Function

Set colorbar norm and ticks in a given range v1 to v2 for data in colorscale.

Vlasiator.set_plot — Method

Configure customized plot.

Private APIs

Vlasiator.AxisUnit — Type

Axis unit type. Currently supported: SI, RE.

Vlasiator.ColorScale — Type

Color scales type for 2D plots. Currently supported: Log, Linear, SymLog.

Vlasiator.PlotArgs — Type

Plotting arguments.

Vlasiator.VMeshInfo — Type

Velocity mesh information.

Base.ndims — Method

ndims(meta) -> Int

Return the dimension of VLSV data.

Base.size — Method

File size in bytes.

Vlasiator.curl — Method

curl(dx::AbstractVector, A::AbstractArray)

Calculate 2nd order cell-centered ∇×A where A is a 4D array of size (3, nx, ny, nz) and dx is a vector of grid intervals in each dimension.

Vlasiator.fillmesh — Method

fillmesh(meta::MetaVLSV, vars; verbose=false) -> celldata, vtkGhostType

Fill the DCCRG mesh with quantity of vars on all refinement levels.

Return arguments

celldata::Vector{Vector{Array}}: data for each variable on each AMR level.

vtkGhostType::Array{UInt8}: cell status (to be completed!).

Vlasiator.get1stcell — Method

Return the first cellid - 1 on mylevel given ncells on this level.

Vlasiator.getObjInfo — Method

Return size and type information for the object.

Vlasiator.getRotationB — Method

getRotationB(B) -> SMatrix

Obtain a rotation matrix with each column being a unit vector which is parallel (v3) and perpendicular (v1,v2) to the magnetic field B. The two perpendicular directions are chosen based on the reference vector of z-axis in the Cartesian coordinates.

Vlasiator.getRotationMatrix — Method

getRotationMatrix(vector, angle)

Creates a rotation matrix that rotates around a unit vector by an angle in radians. References: https://en.wikipedia.org/wiki/Rodrigues'rotationformula https://en.wikipedia.org/wiki/Rotationmatrix#Rotationmatrixfromaxisandangle

Vlasiator.get_axis — Method

Return x and y ranges for 2D.

Vlasiator.getdata2d — Method

Return 2d scalar/vector data. Nonpublic because it won't work with DCCRG AMR.

Vlasiator.getfooter — Method

Return the xml footer of vlsv.

Vlasiator.getindexes — Method

Compute every cell id's x, y and z indexes on the given refinement level (0-based).

Vlasiator.prep2d — Function

Obtain data from meta of var for 2D plotting. Use op to select vector components.

Vlasiator.readmesh — Method

Return mesh related variable.

Vlasiator.readvector — Method

Return vectors of name from the vlsv file with footer opened by fid.

Vlasiator.rotateTensorToVectorZ — Method

rotateTensorToVectorZ(tensor, vector)

Rotates tensor with a rotation matrix that aligns z-axis with vector.

Vlasiator.rotateWithB! — Method

rotateWithB!(T, B)

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB

Vlasiator.rotateWithB — Method

rotateWithB(T, B) -> Matrix

Rotate the tensor T with the 3rd direction aligned with B. See also: rotateWithB!

Vlasiator.save_image — Function

save_image(meta::MetaVLSV, file, vars, data, vtkGhostType, level,
   ascii=false, append=true)

Save data of name vars at AMR level into VTK image file of name file.

Arguments

file::String: output file name.

vars::Vector{String}: variable names to be saved.

data::Vector{Vector}: data for all the variables on each refinement level.

vtkGhostType::Array{UInt8}: array for visibility control.

level::Int: refinement level (0-based).

ascii=false: save output in ASCII or binary format.

append=true: determines whether to append data at the end of file or do in-block writing.

Vlasiator.set_args — Method

Set plot-related arguments of var in axisunit.

Vlasiator.set_lim — Function

Set colormap limits for data.