Internal

Public APIs

Vlasiator.compare — Function

compare(filename1, filename2, tol=1e-4)

Compare if two VLSV files are identical.

Vlasiator.getNearestCellWithVspace — Method

Find the nearest spatial cell with f saved of a given cell id.

Vlasiator.getSliceCellID — Method

getSliceCellID(meta, slicelocation, maxreflevel; xmin=-Inf, xmax=Inf,
   ymin=-Inf, ymax=Inf, zmin=-Inf, zmax=Inf)

Find the cell ids which are needed to plot a 2d cut through of a 3d mesh.

Vlasiator.get_amr_level — Method

get_amr_level(meta::MetaData, cellid)

Return the AMR level of a given cell id.

Vlasiator.get_cell_coordinates — Method

get_cell_coordinates(meta::MetaData, cellid)

Return a given cellid's coordinates.

Vlasiator.get_cell_in_line — Method

get_cell_in_line(meta::MetaData, point1, point2)

Returns cell IDs, distances and coordinates for every cell in a line between two given points. May be improved later with preallocation!

Vlasiator.get_cellid — Method

get_cell_id(meta::MetaData, location)

Return cell ID containing the given spatial location.

Vlasiator.get_max_amr_level — Method

get_max_amr_level(meta::MetaData)

Find the highest refinement level.

Vlasiator.get_variable_derived — Method

get_variable_derived(meta::MetaData, var)

Calculate a derived variable from basic quantities.

Vlasiator.get_velocity_cell_coordinates — Method

get_velocity_cell_coordinates(meta, vcellids, pop="proton")

Returns velocity cells' coordinates of population pop and id vcellids.

Vlasiator.has_name — Method

Check if the XMLElement elem contains a tag with name.

Vlasiator.has_parameter — Method

has_parameter(meta::MetaData, param)

Check if vlsv file contains a parameter.

Vlasiator.has_variable — Method

Check if the VLSV file contains a variable.

Vlasiator.read_meta — Method

read_meta(filename; verbose=false)

Return a struct of MetaData from vlsv file.

Vlasiator.read_parameter — Method

read_parameter(meta::MetaData, param)

Return the parameter value from vlsv file.

Vlasiator.read_variable — Function

read_variable(meta::MetaData, var, sorted=true)

Return variable value from the vlsv file. For DCCRG grid, the variables are sorted by cell ID by default.

Vlasiator.read_variable_info — Method

read_variable_info(meta::MetaData, var)

Return a struct of VarInfo.

Vlasiator.read_variable_select — Function

read_variable_select(meta::MetaData, var, idIn=UInt[])

Read a variable in a collection of cells.

Vlasiator.read_velocity_cells — Method

read_velocity_cells(meta, cellid; pop="proton")

Read velocity cells from a spatial cell of ID cellid, and return a map of velocity cell ids and corresponding value.

Vlasiator.refine_data — Method

refine_data(meta, idlist, data, maxreflevel, normal)

Generate scalar data on the finest refinement level.

Vlasiator.show_variables — Method

Display all variables in the VLSV file.

Private types and methods

Vlasiator.MeshInfo — Type

Mesh size information.

Vlasiator.read_mesh — Method

Return mesh related variable.

Vlasiator.read_prep — Method

Return size and type information for the object.

Vlasiator.read_vector — Method

Return vector data from vlsv file.

Vlasiator.read_xml_footer — Method

Return the xml footer of vlsv.